pytest: Separate config tests
